fix: correct slider desync, current label collision, minCurrent type, button sizing
Addresses final-review findings: the min-current slider no longer
sticks on a rejected value after a failed change (was keyed off a
remember() that never re-ran once evcc's echoed value matched the
pre-change state); the slider's Ladestrom label no longer collides
with the status card's own Ladestrom label (now Mindest-Ladestrom);
LoadpointDto.minCurrent is now Double, matching evcc's actual wire
format and preventing a permanent status failure from one fractional
value; phase/mode buttons now grow instead of clipping if their label
doesn't fit one line. Also: the unlock checkbox now survives
activity recreation (rememberSaveable), and a test now pins the
unconditional-refetch-after-failure invariant the setMode/setPhases/
setMinCurrent refactor depends on.

+++ app/src/main/java/com/example/chargecontrol/network/EvccApi.kt
@@ -14,7 +14,7 @@ data class LoadpointDto(
     val offeredCurrent: Double,
     val connected: Boolean,
     val charging: Boolean,
-    val minCurrent: Int
+    val minCurrent: Double
 )
 
 @Serializable
